Refactor components into their own file; Fix pollInterval functionality
