+++ client/assets/javascripts/App.jsx
@@ -1,9 +1,11 @@
 // The entry point to our client-side app (used by both Rails and Webpack HMR).
 // Essentially renders the top-level view-controller CommentBox component.
 
-var React = require('react/addons');
-var $ = require('jquery');
-import { CommentBox } from './components/CommentBox';
+'use strict';
+
+import $ from 'jquery';
+import React from 'react';
+import CommentBox from './components/CommentBox';
 
 var render = () => {
   if ($("#content").length > 0) {
