+++ client/assets/javascripts/components/Comment.jsx
@@ -1,0 +1,27 @@
+'use strict';
+
+import React from 'react';
+import marked from 'marked';
+
+var Comment = React.createClass({
+  displayName: 'Comment',
+
+  propTypes: {
+    author: React.PropTypes.string.isRequired,
+    text: React.PropTypes.string.isRequired
+  },
+
+  render() {
+    const rawMarkup = marked(this.props.text);
+    return (
+      <div className="comment">
+        <h2 className="commentAuthor foobar">
+          {this.props.author}
+        </h2>
+        <span dangerouslySetInnerHTML={{__html: rawMarkup}} />
+      </div>
+    );
+  }
+});
+
+export default Comment;
