+++ client/assets/javascripts/components/CommentList.jsx
@@ -1,0 +1,31 @@
+'use strict';
+
+import React from 'react';
+import Comment from './Comment';
+
+var CommentList = React.createClass({
+  displayName: 'CommentList',
+
+  propTypes: {
+    comments: React.PropTypes.array
+  },
+
+  render() {
+    var reversedData = this.props.comments.slice(0).reverse();
+    var commentNodes = reversedData.map((comment, index) => {
+      return (
+        // `key` is a React-specific concept and is not mandatory for the
+        // purpose of this tutorial. if you're curious, see more here:
+        // http://facebook.github.io/react/docs/multiple-components.html#dynamic-children
+        <Comment author={comment.author} key={index} text={comment.text}/>
+      );
+    });
+    return (
+      <div className="commentList">
+        {commentNodes}
+      </div>
+    );
+  }
+});
+
+export default CommentList;
